Workaround for game over stopped (#87).
This is my proposal:
This fixes the bug by now, by putting timers instead of on_animation_finished events. But the real problem is that sometimes on_animation_finished is not called, maybe in relation with on_comand_pressed event.
It seems that there are something wrong in this event, or in on_command_pressed, or in the lua object's event management.